+++ packages/@expo/config-plugins/src/utils/plugin-resolver.ts
@@ -11,10 +11,10 @@ export const pluginFileName = 'app.plugin.js';
 /**
  * Resolve the config plugin from a node module or package.
  * If the module or package does not include a config plugin, this function throws a `PluginError`.
- * The resolution is done in following order:
+ * The resolution is done in the following order:
  *   1. Is the reference a relative file path or an import specifier with file path? e.g. `./file.js`, `pkg/file.js` or `@org/pkg/file.js`?
  *     - Resolve the config plugin as-is
- *   2. If the reference a module? e.g. `expo-font`
+ *   2. Is the reference a module? e.g. `expo-font`
  *     - Resolve the root `app.plugin.js` file within the module, e.g. `expo-font/app.plugin.js`
  *   3. Does the module have a valid config plugin in the `main` field?
  *     - Resolve the `main` entry point as config plugin
@@ -25,7 +25,7 @@ export function resolvePluginForModule(
 ): { filePath: string; isPluginFile: boolean } {
   if (moduleNameIsDirectFileReference(pluginReference)) {
     // Only resolve `./file.js`, `package/file.js`, `@org/package/file.js`
-    const pluginScriptFile = resolveFrom.silent(projectRoot, pluginReference);
+    const pluginScriptFile = exportsAwareSilentResolve(projectRoot, pluginReference);
     if (pluginScriptFile) {
       return {
         // NOTE(cedric): `path.sep` is required here, we are resolving the absolute path, not the plugin reference
@@ -35,20 +35,19 @@ export function resolvePluginForModule(
     }
   } else if (moduleNameIsPackageReference(pluginReference)) {
     // Only resolve `package -> package/app.plugin.js`, `@org/package -> @org/package/app.plugin.js`
-    const pluginPackageFile = resolveFrom.silent(
+    const pluginPackageFile = exportsAwareSilentResolve(
       projectRoot,
       `${pluginReference}/${pluginFileName}`
     );
     if (pluginPackageFile && fileExists(pluginPackageFile)) {
       return { isPluginFile: true, filePath: pluginPackageFile };
     }
-    // Try to resole the `main` entry as config plugin
+    // Try to resolve the `main` entry as config plugin
     const packageMainEntry = resolveFrom.silent(projectRoot, pluginReference);
     if (packageMainEntry) {
       return { isPluginFile: false, filePath: packageMainEntry };
     }
   }
-
   throw new PluginError(
     `Failed to resolve plugin for module "${pluginReference}" relative to "${projectRoot}". Do you have node modules installed?`,
     'PLUGIN_NOT_FOUND'
@@ -75,6 +74,26 @@ export function moduleNameIsDirectFileReference(name: string): boolean {
   // Regular packages should be considered direct reference if they have more than one slash.
   return slashCount > 1;
 }
+
+const exportsAwareSilentResolve = (projectRoot: string, pluginReference: string): string | null => {
+  try {
+    return resolveFrom(projectRoot, pluginReference);
+  } catch (error) {
+    if (
+      error &&
+      typeof error === 'object' &&
+      'code' in error &&
+      error.code === 'ERR_PACKAGE_PATH_NOT_EXPORTED'
+    ) {
+      throw new PluginError(
+        `Failed to resolve plugin for module "${pluginReference}" relative to "${projectRoot}". The author of "${pluginReference}" should add the "app.plugin.js" entry to the "exports" field of that module's package.json.`,
+        'PLUGIN_NOT_FOUND',
+        error as unknown as Error
+      );
+    }
+    return null;
+  }
+};
 
 export function moduleNameIsPackageReference(name: string): boolean {
   const slashCount = name.split('/')?.length;